+++ packages/runtime/container-runtime/src/opLifecycle/index.ts
@@ -7,7 +7,6 @@ export {
 	BatchId,
 	BatchManager,
 	BatchSequenceNumbers,
-	estimateSocketSize,
 	getEffectiveBatchId,
 	generateBatchId,
 	IBatchManagerOptions,
@@ -15,6 +14,7 @@ export {
 export {
 	LocalBatch,
 	LocalBatchMessage,
+	LocalEmptyBatchPlaceholder,
 	OutboundBatch,
 	OutboundBatchMessage,
 	OutboundSingletonBatch,
@@ -26,7 +26,12 @@ export {
 	serializeOp,
 	ensureContentsDeserialized,
 } from "./opSerialization.js";
-export { Outbox, getLongStack } from "./outbox.js";
+export {
+	estimateSocketSize,
+	localBatchToOutboundBatch,
+	Outbox,
+	getLongStack,
+} from "./outbox.js";
 export { OpCompressor } from "./opCompressor.js";
 export { OpDecompressor } from "./opDecompressor.js";
 export { OpSplitter, splitOp, isChunkedMessage } from "./opSplitter.js";