+++ packages/runtime/container-runtime/src/test/opLifecycle/outbox.spec.ts
@@ -39,6 +39,9 @@ import {
 	Outbox,
 	type LocalBatchMessage,
 	type OutboundBatch,
+	localBatchToOutboundBatch,
+	serializeOp,
+	type LocalEmptyBatchPlaceholder,
 } from "../../opLifecycle/index.js";
 import {
 	PendingMessageResubmitData,
@@ -46,9 +49,18 @@ import {
 	type IPendingMessage,
 } from "../../pendingStateManager.js";
 
-function typeFromBatchedOp(op: LocalBatchMessage): string {
-	assert(op.serializedOp !== undefined, "PRECONDITION: serializedOp is undefined");
-	return (JSON.parse(op.serializedOp) as LocalContainerRuntimeMessage).type;
+function typeFromBatchedOp(message: LocalBatchMessage): string {
+	assert(message.runtimeOp !== undefined, "PRECONDITION: runtimeOp is undefined");
+	return message.runtimeOp.type;
+}
+
+// Make a mock op with distinguishable contents
+function op(data: string): LocalContainerRuntimeMessage {
+	// eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+	return {
+		type: ContainerMessageType.FluidDataStoreOp,
+		contents: data as unknown,
+	} as LocalContainerRuntimeMessage;
 }
 
 describe("Outbox", () => {
@@ -138,23 +150,41 @@ describe("Outbox", () => {
 			clientSequenceNumber: number | undefined,
 		): void => {
 			for (const {
-				serializedOp: content,
+				runtimeOp,
 				referenceSequenceNumber,
 				metadata: opMetadata,
 				localOpMetadata,
 			} of batch)
 				state.pendingOpContents.push({
-					content,
+					runtimeOp,
 					referenceSequenceNumber,
 					opMetadata,
 					localOpMetadata,
 					batchStartCsn: clientSequenceNumber ?? -1,
 				});
 		},
+		onFlushEmptyBatch: (
+			{
+				metadata: opMetadata,
+				localOpMetadata,
+				referenceSequenceNumber,
+			}: LocalEmptyBatchPlaceholder,
+			clientSequenceNumber: number | undefined,
+		) =>
+			state.pendingOpContents.push({
+				referenceSequenceNumber,
+				opMetadata,
+				localOpMetadata,
+				batchStartCsn: clientSequenceNumber,
+			}),
 	});
 
-	const createMessage = (type: ContainerMessageType, contents: string): LocalBatchMessage => ({
-		serializedOp: JSON.stringify({ type, contents }),
+	const createMessage = (
+		type: ContainerMessageType,
+		fakeContents: string,
+	): LocalBatchMessage => ({
+		// eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+		runtimeOp: { type, contents: fakeContents as unknown } as LocalContainerRuntimeMessage,
 		metadata: undefined,
 		localOpMetadata: {},
 		referenceSequenceNumber: Number.POSITIVE_INFINITY,
@@ -164,7 +194,7 @@ describe("Outbox", () => {
 		message: LocalBatchMessage | OutboundBatchMessage,
 		batchMarker: boolean | undefined = undefined,
 	): IBatchMessage => ({
-		contents: "serializedOp" in message ? message.serializedOp : message.contents,
+		contents: "runtimeOp" in message ? serializeOp(message.runtimeOp) : message.contents,
 		metadata:
 			batchMarker === undefined
 				? message.metadata
@@ -174,7 +204,7 @@ describe("Outbox", () => {
 	});
 
 	// Also converts to an OutboundBatchMessage
-	const addBatchMetadata = (messages: LocalBatchMessage[]): OutboundBatchMessage[] => {
+	const addBatchMetadata = (messages: OutboundBatchMessage[]): void => {
 		if (messages.length > 1) {
 			messages[0].metadata = {
 				...messages[0].metadata,
@@ -185,21 +215,18 @@ describe("Outbox", () => {
 				batch: false,
 			};
 		}
-
-		return messages.map<OutboundBatchMessage>(({ serializedOp, ...message }) => ({
-			contents: serializedOp,
-			...message,
-		}));
 	};
-	const toOutboundBatch = (messages: LocalBatchMessage[]): OutboundBatch => ({
-		messages: addBatchMetadata(messages),
-		contentSizeInBytes: messages
-			.map((message) => message.serializedOp?.length ?? 0)
-			.reduce((a, b) => a + b, 0),
-		referenceSequenceNumber:
-			messages.length === 0 ? undefined : messages[0].referenceSequenceNumber,
-		hasReentrantOps: false,
-	});
+	const toOutboundBatch = (messages: LocalBatchMessage[]): OutboundBatch => {
+		const outbound = localBatchToOutboundBatch({
+			messages,
+			referenceSequenceNumber:
+				messages.length === 0 ? undefined : messages[0].referenceSequenceNumber,
+			hasReentrantOps: false,
+		});
+
+		addBatchMetadata(outbound.messages);
+		return outbound;
+	};
 
 	const DefaultCompressionOptions = {
 		minimumBatchSizeInBytes: Number.POSITIVE_INFINITY,
@@ -274,6 +301,34 @@ describe("Outbox", () => {
 		mockLogger.clear();
 	});
 
+	it("localBatchToOutboundBatch", () => {
+		const localMessages: LocalBatchMessage[] = [
+			{ runtimeOp: op("hello"), referenceSequenceNumber: 4 },
+			{ runtimeOp: op("world"), referenceSequenceNumber: 4 },
+			{ runtimeOp: op("!"), referenceSequenceNumber: 4 },
+		];
+		const localBatch = {
+			messages: localMessages,
+			referenceSequenceNumber: localMessages[0].referenceSequenceNumber,
+			hasReentrantOps: false,
+		};
+		const outboundBatch = localBatchToOutboundBatch(localBatch);
+
+		// Check that contentSizeInBytes and messages' contents are set propertly
+		assert.equal(
+			outboundBatch.contentSizeInBytes,
+			JSON.stringify(op("")).length * 3 + "helloworld!".length,
+			"contentSizeInBytes is incorrect",
+		);
+		assert.equal(outboundBatch.messages.length, 3);
+		assert.deepEqual(
+			localMessages.map((m) => m.runtimeOp),
+			// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-non-null-assertion
+			outboundBatch.messages.map((m) => JSON.parse(m.contents!)),
+			"Serialized contents do not match",
+		);
+	});
+
 	it("Sending batches", () => {
 		const outbox = getOutbox({ context: getMockContext() });
 		const messages = [
@@ -326,7 +381,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -450,7 +505,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			messages.map<Partial<IPendingMessage>>((message, i) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -499,7 +554,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -564,7 +619,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -573,12 +628,15 @@ describe("Outbox", () => {
 		);
 	});
 
-	it("Compress only if the batch is larger than the configured limit", () => {
+	it("Does not compress if the batch is smaller than the configured limit", () => {
 		const outbox = getOutbox({
 			context: getMockContext(),
-			maxBatchSize: 1,
+			maxBatchSize: 1024,
+			opGroupingConfig: {
+				groupedBatchingEnabled: true, // Required for compression
+			},
 			compressionOptions: {
-				minimumBatchSizeInBytes: 1024,
+				minimumBatchSizeInBytes: 512,
 				compressionAlgorithm: CompressionAlgorithms.lz4,
 			},
 		});
@@ -597,22 +655,15 @@ describe("Outbox", () => {
 
 		outbox.flush();
 
-		assert.equal(state.opsSubmitted, messages.length);
+		assert.equal(
+			state.opsSubmitted,
+			2,
+			"Expected 2 ops to be submitted, on ID Allocation and one for the grouped batch",
+		);
 		assert.equal(state.batchesSubmitted.length, 2);
 		assert.equal(state.individualOpsSubmitted.length, 0);
 		assert.equal(state.deltaManagerFlushCalls, 0);
 		assert.deepEqual(state.batchesCompressed, []);
-		assert.deepEqual(
-			state.batchesSubmitted.map((x) => x.messages),
-			[
-				[toSubmittedMessage(messages[2])],
-				[
-					toSubmittedMessage(messages[0], true),
-					toSubmittedMessage(messages[1]),
-					toSubmittedMessage(messages[3], false),
-				],
-			],
-		);
 
 		// Note the expected CSN here is fixed to the batch's starting CSN
 		const expectedMessageOrderWithCsn = [
@@ -626,7 +677,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -658,7 +709,13 @@ describe("Outbox", () => {
 		outbox.submit(messages[1]);
 		outbox.submit(messages[2]);
 
-		assert.throws(() => outbox.flush());
+		assert.throws(
+			() => outbox.flush(),
+			(e: Error) =>
+				"dataProcessingCodepath" in e &&
+				e.dataProcessingCodepath === "CompressionInsufficient",
+			"Expected 'CompressionInsufficient' error",
+		);
 		// The batch is compressed
 		assert.deepEqual(state.batchesCompressed, [
 			opGroupingManager.groupBatch(toOutboundBatch(messages)),
@@ -670,7 +727,7 @@ describe("Outbox", () => {
 	it("Chunks when compression is enabled, compressed batch is larger than the threshold and chunking is enabled", () => {
 		const outbox = getOutbox({
 			context: getMockContext(),
-			maxBatchSize: 1,
+			maxBatchSize: 1024,
 			compressionOptions: {
 				minimumBatchSizeInBytes: 1,
 				compressionAlgorithm: CompressionAlgorithms.lz4,
@@ -722,7 +779,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -734,7 +791,7 @@ describe("Outbox", () => {
 	it("Does not chunk when compression and grouping are enabled, compressed batch is smaller than the threshold and chunking is enabled", () => {
 		const outbox = getOutbox({
 			context: getMockContext(),
-			maxBatchSize: 1,
+			maxBatchSize: 1024,
 			compressionOptions: {
 				minimumBatchSizeInBytes: 1,
 				compressionAlgorithm: CompressionAlgorithms.lz4,
@@ -836,7 +893,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			rawMessagesInFlushOrder.map((message, i) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,
@@ -851,7 +908,7 @@ describe("Outbox", () => {
 		]);
 	});
 
-	for (const ops of [
+	for (const messages of [
 		[
 			{
 				...createMessage(ContainerMessageType.IdAllocation, "0"),
@@ -886,21 +943,21 @@ describe("Outbox", () => {
 				context: getMockContext(),
 				flushPartialBatches: true,
 			});
-			for (const op of ops) {
-				currentSeqNumbers.referenceSequenceNumber = op.referenceSequenceNumber;
-				if (typeFromBatchedOp(op) === ContainerMessageType.IdAllocation) {
-					outbox.submitIdAllocation(op);
+			for (const message of messages) {
+				currentSeqNumbers.referenceSequenceNumber = message.referenceSequenceNumber;
+				if (typeFromBatchedOp(message) === ContainerMessageType.IdAllocation) {
+					outbox.submitIdAllocation(message);
 				} else {
-					outbox.submit(op);
+					outbox.submit(message);
 				}
 			}
 
-			assert.equal(state.opsSubmitted, ops.length - 1);
+			assert.equal(state.opsSubmitted, messages.length - 1);
 			assert.equal(state.individualOpsSubmitted.length, 0);
 			assert.equal(state.batchesSubmitted.length, 1);
 			assert.deepEqual(
 				state.batchesSubmitted.map((x) => x.messages),
-				[[toSubmittedMessage(ops[0]), toSubmittedMessage(ops[1])]],
+				[[toSubmittedMessage(messages[0]), toSubmittedMessage(messages[1])]],
 			);
 
 			mockLogger.assertMatch([
@@ -1009,7 +1066,7 @@ describe("Outbox", () => {
 		assert.deepEqual(
 			state.pendingOpContents,
 			expectedMessageOrderWithCsn.map<Partial<IPendingMessage>>(([message, csn]) => ({
-				content: message.serializedOp,
+				runtimeOp: message.runtimeOp,
 				referenceSequenceNumber: message.referenceSequenceNumber,
 				localOpMetadata: message.localOpMetadata,
 				opMetadata: message.metadata,